Initial switch to RSA for api keys

+++ src/main/java/sweng/penelope/controllers/ApiKeyController.java
@@ -1,7 +1,11 @@
 package sweng.penelope.controllers;
 
+import java.nio.file.FileSystemException;
+import java.security.KeyPair;
 import java.security.NoSuchAlgorithmException;
+import java.security.PublicKey;
 import java.security.SecureRandom;
+import java.util.Base64;
 import java.util.Optional;
 import java.util.Set;
 
@@ -15,10 +19,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import sweng.penelope.Responses;
+import sweng.penelope.auth.RSAUtils;
 import sweng.penelope.entities.ApiKey;
 import sweng.penelope.entities.Campus;
 import sweng.penelope.repositories.ApiKeyRepository;
 import sweng.penelope.repositories.CampusRepository;
+import sweng.penelope.services.StorageService;
 
 @Controller
 @RequestMapping(path = "/api/apikeys")
@@ -27,10 +33,11 @@ public class ApiKeyController {
     private ApiKeyRepository apiKeyRepository;
     @Autowired
     private CampusRepository campusRepository;
+    @Autowired
+    private StorageService storageService;
 
     private static final String[] CHARS = "abcdefghijklmnoprstuvwxyz0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ".split("");
-    private static final int PUBLIC_KEY_LENGTH = 10;
-    private static final int PRIVATE_KEY_LENGTH = 20;
+    private static final int IDENTITY_LENGTH = 10;
 
     private String generateString(int length, SecureRandom secureRandom) {
         StringBuilder builder = new StringBuilder();
@@ -50,55 +57,63 @@ public class ApiKeyController {
         try {
             SecureRandom secureRandom = SecureRandom.getInstanceStrong();
 
-            String publicKey;
-            String privateKey;
+            String identity;
 
             while (true) {
-                publicKey = generateString(PUBLIC_KEY_LENGTH, secureRandom);
+                identity = generateString(IDENTITY_LENGTH, secureRandom);
 
-                if (!apiKeyRepository.findById(publicKey).isPresent())
+                if (!apiKeyRepository.findById(identity).isPresent())
                     break;
             }
 
-            privateKey = generateString(PRIVATE_KEY_LENGTH, secureRandom);
+            KeyPair keyPair = RSAUtils.generateKeys();
 
-            apiKey.setKey(publicKey);
-            apiKey.setSecret(privateKey);
+            if (!storageService.storeKey(keyPair.getPrivate(), identity))
+                throw new FileSystemException("Could not store key");
+
+            PublicKey publicKey = keyPair.getPublic();
+            String publicKeyBase64 = Base64.getEncoder().encodeToString(publicKey.getEncoded());
+
+            apiKey.setIdentity(identity);
 
             apiKeyRepository.save(apiKey);
 
-            return ResponseEntity.ok().body(String.format("%s:%s%n", publicKey, privateKey));
-        } catch (NoSuchAlgorithmException e) {
+            return ResponseEntity.ok().body(String.format("%s:%s%n", identity, publicKeyBase64));
+        } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
         }
     }
 
     @DeleteMapping(path = "/remove")
-    public ResponseEntity<String> removeApiKey(@RequestParam String publicKey) {
-        Optional<ApiKey> requestedKeyToBeDeleted = apiKeyRepository.findById(publicKey);
+    public ResponseEntity<String> removeApiKey(@RequestParam String identity) {
+        Optional<ApiKey> requestedKeyToBeDeleted = apiKeyRepository.findById(identity);
 
         if (requestedKeyToBeDeleted.isPresent()) {
             ApiKey keyToBeDeleted = requestedKeyToBeDeleted.get();
 
-            apiKeyRepository.delete(keyToBeDeleted);
+            if (storageService.removeKey(keyToBeDeleted.getIdentity())) {
+                apiKeyRepository.delete(keyToBeDeleted);
 
-            return ResponseEntity.ok().body(String.format("Key %s deleted.%n", publicKey));
+                return ResponseEntity.ok().body(String.format("Key %s deleted.%n", identity));
+            } else {
+                return ResponseEntity.internalServerError().build();
+            }
         }
 
         Responses responses = new Responses();
-        return responses.notFound(String.format("Key %s was not found. Nothing to do here...%n", publicKey));
+        return responses.notFound(String.format("Key %s was not found. Nothing to do here...%n", identity));
     }
 
     @PatchMapping(path = "/addCampus")
-    public ResponseEntity<String> addCampusToKey(@RequestParam Long campusId, @RequestParam String publicKey) {
-        Optional<ApiKey> requestKey = apiKeyRepository.findById(publicKey);
+    public ResponseEntity<String> addCampusToKey(@RequestParam Long campusId, @RequestParam String identity) {
+        Optional<ApiKey> requestKey = apiKeyRepository.findById(identity);
         Optional<Campus> requestCampus = campusRepository.findById(campusId);
 
         Responses responses = new Responses();
 
         if (requestKey.isEmpty())
-            return responses.notFound(String.format("Public key %s not found. Nothing to do here...%n", publicKey));
+            return responses.notFound(String.format("Public key %s not found. Nothing to do here...%n", identity));
 
         if (requestCampus.isEmpty())
             return responses.notFound(String.format("Campus %d not found. Nothing to do here...%n", campusId));
@@ -113,18 +128,18 @@ public class ApiKeyController {
 
         apiKeyRepository.save(apiKey);
 
-        return ResponseEntity.ok().body(String.format("Campus %d added to key %s.%n", campusId, publicKey));
+        return ResponseEntity.ok().body(String.format("Campus %d added to key %s.%n", campusId, identity));
     }
 
     @PatchMapping(path = "/removeCampus")
-    public ResponseEntity<String> removeCampusFromKey(@RequestParam Long campusId, @RequestParam String publicKey) {
-        Optional<ApiKey> requestKey = apiKeyRepository.findById(publicKey);
+    public ResponseEntity<String> removeCampusFromKey(@RequestParam Long campusId, @RequestParam String identity) {
+        Optional<ApiKey> requestKey = apiKeyRepository.findById(identity);
         Optional<Campus> requestCampus = campusRepository.findById(campusId);
 
         Responses responses = new Responses();
 
         if (requestKey.isEmpty())
-            return responses.notFound(String.format("Public key %s not found. Nothing to do here...%n", publicKey));
+            return responses.notFound(String.format("Public key %s not found. Nothing to do here...%n", identity));
 
         if (requestCampus.isEmpty())
             return responses.notFound(String.format("Campus %d not found. Nothing to do here...%n", campusId));
@@ -141,9 +156,9 @@ public class ApiKeyController {
 
             apiKeyRepository.save(apiKey);
 
-            return ResponseEntity.ok().body(String.format("Campus %d removed from key %s.%n", campusId, publicKey));
+            return ResponseEntity.ok().body(String.format("Campus %d removed from key %s.%n", campusId, identity));
         }
 
-        return responses.notFound(String.format("Key %s does not have rights on campus %d.%n", publicKey, campusId));
+        return responses.notFound(String.format("Key %s does not have rights on campus %d.%n", identity, campusId));
     }
 }
